Avoid using sys.argv for internal parser by default
Passing args into the parse_args function adds flexibility.

When 'None' is passed as args, the parser will read sys.argv,
allowing the user to pass in command line args when calling
'python hubconf.py'.

When [] is pased in as args, the parser relies only on defaults,
which is safe for cases where the model is imported somewhere
else and might be used with a different arg parser.

Since the API for get_module, train, eval don't pass 'args' values,
make the default args [] for compatibility and override with None
when needed.